+++ src/components/custom-field-input-currency/custom-field-input-currency.jsx
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import React, { useState, useRef, useEffect } from 'react';
+import React, { forwardRef, useImperativeHandle, useState, useRef, useEffect } from 'react';
 import CustomFieldInputText from '../custom-field-input-text/custom-field-input-text.jsx';
 import CustomFieldInputNumber from '../custom-field-input-number/custom-field-input-number.jsx';
 import currencyCodeType from './currency-code-type.js';
@@ -39,7 +39,8 @@ function formatValue(unitValue, currencyCode) {
   }).format(unitValue);
 }
 
-export default function CustomFieldInputCurrency(props) {
+const CustomFieldInputCurrency = forwardRef(function CustomFieldInputCurrency(props, ref) {
+  const componentRef = useRef(null);
   const [input, setInput] = useState(subunitToUnit(props.value, props.currencyCode));
   const [isEditing, setIsEditing] = useState(false);
   const [isFocused, setIsFocused] = useState(false);
@@ -73,6 +74,12 @@ export default function CustomFieldInputCurrency(props) {
     }
   });
 
+  useImperativeHandle(ref, () => ({
+    value: () => {
+      return [parseInt(componentRef.current.value().replace(/\D/g, ''), 10), props.currencyCode];
+    },
+  }));
+
   const sharedProps = {
     className: props.className,
     disabled: props.disabled,
@@ -104,10 +111,11 @@ export default function CustomFieldInputCurrency(props) {
       defaultValue={formattedNumber}
       errorText={props.errorText}
       onFocus={handleOnFocus}
+      ref={componentRef}
       type="text"
     />
   );
-}
+});
 
 CustomFieldInputCurrency.propTypes = {
   className: PropTypes.string,
@@ -135,3 +143,5 @@ CustomFieldInputCurrency.defaultProps = {
   required: false,
   value: undefined,
 };
+
+export default CustomFieldInputCurrency;